:recycle: api: Move mail-check to src
